observability/math-otlp/math-service/src/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi_class==3.7.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
mpmath==1.3.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==25.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
protobuf==6.32.0
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.5
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.3
zipp==3.23.0
